Add files via upload
Program which simulates an online shopping cart. This provides users with a basic menu and
gives them the option to add items to their cart, remove items, view the current total cost of
their cart, or ‘checkout’ and exit the program. Also uses defensive programming techniques to
ensure the user enters valid inputs.